[kernel] Consolidate directhd.h and hdreg.h
